+++ src/main/java/com/dolgikh/scriptorium/repositories/BooksRepository.java
@@ -3,5 +3,5 @@ package com.dolgikh.scriptorium.repositories;
 import com.dolgikh.scriptorium.models.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BooksRepository extends JpaRepository<Book, Integer> {
+public interface BooksRepository extends JpaRepository<Book, Long> {
 }
