+++ src/main/java/com/dolgikh/scriptorium/services/AuthorsService.java
@@ -25,7 +25,7 @@ public class AuthorsService {
         return authorsRepository.findAll();
     }
 
-    public Author findOne(int id) {
+    public Author findOne(long id) {
         Optional<Author> author = authorsRepository.findById(id);
 
         if (author.isEmpty())
@@ -34,7 +34,7 @@ public class AuthorsService {
         return author.get();
     }
 
-    public List<Book> findBooksOfAuthor(int id) {
+    public List<Book> findBooksOfAuthor(long id) {
         return findOne(id).getBooks();
     }
 
@@ -44,7 +44,7 @@ public class AuthorsService {
     }
 
     @Transactional
-    public void update(Author author, int id) {
+    public void update(Author author, long id) {
         if (authorsRepository.findById(id).isEmpty())
             throw new AuthorNotFoundException(id);
 
@@ -53,7 +53,7 @@ public class AuthorsService {
     }
 
     @Transactional
-    public void delete(int id) {
+    public void delete(long id) {
         if (authorsRepository.findById(id).isEmpty())
             throw new AuthorNotFoundException(id);
