+++ src/main/java/com/dolgikh/scriptorium/util/exceptions/notfoundexceptions/AuthorNotFoundException.java
@@ -1,7 +1,7 @@
 package com.dolgikh.scriptorium.util.exceptions.notfoundexceptions;
 
 public class AuthorNotFoundException extends ResourceNotFoundException {
-    public AuthorNotFoundException(int id) {
+    public AuthorNotFoundException(long id) {
         super("Author", id);
     }
 }
