+++ src/main/java/com/dolgikh/scriptorium/util/exceptions/notfoundexceptions/ReviewNotFoundException.java
@@ -1,7 +1,7 @@
 package com.dolgikh.scriptorium.util.exceptions.notfoundexceptions;
 
 public class ReviewNotFoundException extends ResourceNotFoundException {
-    public ReviewNotFoundException(int id) {
+    public ReviewNotFoundException(long id) {
         super("Review", id);
     }
 }
